Renames files, adds Header
